Add ability to exclude files from rsync
Add the ability to prevent certain files or directories from being
sent to the backup server via rsync. Any file under a specified
directory will also be excluded. Entries may be specified in the
barman config file as a comma-separated string, and need not
actually exist.

This is useful if you have tablespaces or extensions that you don't
want to back up via rsync on every run, or to prevent secrets like
encryption keys from being included in the backup.